pyproject.toml
[project]
name = "merysu-sparkle-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "HaroldR23",email = "ahsro99@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.122.0,<0.123.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "resend (>=2.19.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
